WebToggleButton proper painter support Buttons painter interfaces restructurization and simplifications All painting code moved into WebAbstractButtonPainter
