+++ modules/ui/src/com/alee/laf/button/AdaptiveButtonPainter.java
@@ -4,7 +4,6 @@ import com.alee.extended.painter.AdaptivePainter;
 import com.alee.extended.painter.Painter;
 
 import javax.swing.*;
-import java.awt.*;
 
 /**
  * Simple ButtonPainter adapter class.
@@ -23,32 +22,5 @@ public class AdaptiveButtonPainter<E extends JButton, U extends WebButtonUI> ext
     public AdaptiveButtonPainter ( final Painter painter )
     {
         super ( painter );
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setIconRect ( final Rectangle rect )
-    {
-        // Ignore this method in adaptive class
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setTextRect ( final Rectangle rect )
-    {
-        // Ignore this method in adaptive class
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setViewRect ( final Rectangle rect )
-    {
-        // Ignore this method in adaptive class
     }
 }
