+++ modules/ui/src/com/alee/laf/button/ButtonPainter.java
@@ -1,10 +1,8 @@
 package com.alee.laf.button;
 
-import com.alee.extended.painter.Painter;
 import com.alee.extended.painter.SpecificPainter;
 
 import javax.swing.*;
-import java.awt.*;
 
 /**
  * Base interface for JButton component painters.
@@ -12,26 +10,6 @@ import java.awt.*;
  * @author Mikle Garin
  */
 
-public interface ButtonPainter<E extends JButton, U extends WebButtonUI> extends Painter<E, U>, SpecificPainter
+public interface ButtonPainter<E extends JButton, U extends WebButtonUI> extends AbstractButtonPainter<E, U>, SpecificPainter
 {
-    /**
-     * Sets button view bounds.
-     *
-     * @param rect button view bounds
-     */
-    public void setViewRect ( Rectangle rect );
-
-    /**
-     * Sets button text bounds.
-     *
-     * @param rect button text bounds
-     */
-    public void setTextRect ( Rectangle rect );
-
-    /**
-     * Sets button icon bounds.
-     *
-     * @param rect button icon bounds
-     */
-    public void setIconRect ( Rectangle rect );
 }
