+++ modules/ui/src/com/alee/laf/button/ToggleButtonPainter.java
@@ -1,9 +1,16 @@
 package com.alee.laf.button;
 
+import com.alee.extended.painter.SpecificPainter;
+
+import javax.swing.*;
+
 /**
+ * Base interface for JButton component painters.
+ *
  * @author Mikle Garin
  */
 
-public interface ToggleButtonPainter extends ButtonPainter
+public interface ToggleButtonPainter<E extends JToggleButton, U extends WebToggleButtonUI>
+        extends AbstractButtonPainter<E, U>, SpecificPainter
 {
 }
