+++ modules/ui/src/com/alee/laf/button/WebToggleButtonUI.java
@@ -17,26 +17,175 @@
 
 package com.alee.laf.button;
 
+import com.alee.extended.painter.Painter;
+import com.alee.extended.painter.PainterSupport;
+import com.alee.managers.style.StyleManager;
+import com.alee.utils.SwingUtils;
+import com.alee.utils.laf.ShapeProvider;
+import com.alee.utils.laf.Styleable;
+import com.alee.utils.swing.DataRunnable;
+
 import javax.swing.*;
 import javax.swing.plaf.ComponentUI;
+import javax.swing.plaf.basic.BasicToggleButtonUI;
+import java.awt.*;
 
 /**
  * @author Mikle Garin
  */
 
-public class WebToggleButtonUI extends WebButtonUI
+public class WebToggleButtonUI extends BasicToggleButtonUI implements Styleable, ShapeProvider, SwingConstants
 {
+    /**
+     * Toggle button property prefix.
+     */
     private final static String propertyPrefix = "ToggleButton" + ".";
 
-    @SuppressWarnings ("UnusedParameters")
+    /**
+     * Component painter.
+     */
+    protected ToggleButtonPainter painter;
+
+    /**
+     * Runtime variables.
+     */
+    protected String styleId = null;
+    protected AbstractButton button = null;
+
+    /**
+     * Returns an instance of the WebToggleButtonUI for the specified component.
+     * This tricky method is used by UIManager to create component UIs when needed.
+     *
+     * @param c component that will use UI instance
+     * @return instance of the WebToggleButtonUI
+     */
+    @SuppressWarnings ( "UnusedParameters" )
     public static ComponentUI createUI ( final JComponent c )
     {
         return new WebToggleButtonUI ();
     }
 
+    /**
+     * Installs UI in the specified component.
+     *
+     * @param c component for this UI
+     */
+    @Override
+    public void installUI ( final JComponent c )
+    {
+        super.installUI ( c );
+
+        // Saving button reference
+        button = ( AbstractButton ) c;
+
+        // Applying skin
+        StyleManager.applySkin ( button );
+    }
+
+    /**
+     * Uninstalls UI from the specified component.
+     *
+     * @param c component with this UI
+     */
+    @Override
+    public void uninstallUI ( final JComponent c )
+    {
+        // Uninstalling applied skin
+        StyleManager.removeSkin ( button );
+
+        // Removing button reference
+        button = null;
+
+        super.uninstallUI ( c );
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     protected String getPropertyPrefix ()
     {
         return propertyPrefix;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String getStyleId ()
+    {
+        return styleId;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void setStyleId ( final String id )
+    {
+        this.styleId = id;
+        StyleManager.applySkin ( button );
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Shape provideShape ()
+    {
+        return PainterSupport.getShape ( button, painter );
+    }
+
+    /**
+     * Returns toggle button painter.
+     *
+     * @return toggle button painter
+     */
+    public Painter getPainter ()
+    {
+        return PainterSupport.getAdaptedPainter ( painter );
+    }
+
+    /**
+     * Sets toggle button painter.
+     * Pass null to remove toggle button painter.
+     *
+     * @param painter new toggle button painter
+     */
+    public void setPainter ( final Painter painter )
+    {
+        PainterSupport.setPainter ( button, new DataRunnable<ToggleButtonPainter> ()
+        {
+            @Override
+            public void run ( final ToggleButtonPainter newPainter )
+            {
+                WebToggleButtonUI.this.painter = newPainter;
+            }
+        }, this.painter, painter, ToggleButtonPainter.class, AdaptiveToggleButtonPainter.class );
+    }
+
+    /**
+     * Paints toggle button.
+     *
+     * @param g graphics
+     * @param c component
+     */
+    @Override
+    public void paint ( final Graphics g, final JComponent c )
+    {
+        if ( painter != null )
+        {
+            // Painting button
+            painter.paint ( ( Graphics2D ) g, SwingUtils.size ( c ), c, this );
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Dimension getPreferredSize ( final JComponent c )
+    {
+        return PainterSupport.getPreferredSize ( c, super.getPreferredSize ( c ), painter );
     }
 }
