+++ modules/ui/src/com/alee/managers/style/skin/web/WebToggleButtonPainter.java
@@ -1,0 +1,15 @@
+package com.alee.managers.style.skin.web;
+
+import com.alee.laf.button.ToggleButtonPainter;
+import com.alee.laf.button.WebToggleButtonUI;
+
+import javax.swing.*;
+
+/**
+ * @author Mikle Garin
+ */
+
+public class WebToggleButtonPainter<E extends JToggleButton, U extends WebToggleButtonUI> extends WebAbstractButtonPainter<E, U>
+        implements ToggleButtonPainter<E, U>
+{
+}
